Added RepositoryService tests

+++ src/main/java/com/undefinedlabs/scope/service/RestaurantService.java
@@ -5,8 +5,11 @@ import com.undefinedlabs.scope.model.Restaurant;
 import com.undefinedlabs.scope.repository.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
+import javax.validation.constraints.NotNull;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -23,19 +26,40 @@ public class RestaurantService {
         return this.repository.findAll();
     }
 
-    public Restaurant getRestaurantById(final String id){
+    public Restaurant getRestaurantById(@NotNull final String id){
+        if (StringUtils.isEmpty(id)){
+            throw new IllegalArgumentException("id cannot be empty");
+        }
+
         return this.repository.findById(UUID.fromString(id)).orElseThrow(() -> new RestaurantNotFoundException("Restaurant " + id + " not found"));
     }
 
-    public List<Restaurant> getRestaurantsByName(final String name) {
+    public List<Restaurant> getRestaurantsByName(@NotNull  final String name) {
+        if(StringUtils.isEmpty(name)){
+            throw new IllegalArgumentException("name cannot be empty");
+        }
+
         return this.repository.findByNameContains(name);
     }
 
-    public Restaurant createRestaurant(final Restaurant restaurant) {
+    public Restaurant createRestaurant(@NotNull final Restaurant restaurant) {
+        if(restaurant == null) {
+            throw new IllegalArgumentException("restaurant cannot be null");
+        }
+
         return this.repository.save(restaurant);
     }
 
-    public Restaurant updateRestaurant(final String id, final Restaurant restaurant) {
+    public Restaurant updateRestaurant(@NotNull final String id, @NotNull final Restaurant restaurant) {
+        if(StringUtils.isEmpty(id)) {
+            throw new IllegalArgumentException("id cannot be empty");
+        }
+
+        if(restaurant == null) {
+            throw new IllegalArgumentException("restaurant cannot be null");
+        }
+
+
         final Restaurant toUpdate = this.repository.getOne(UUID.fromString(id));
 
         toUpdate.setName(restaurant.getName());
@@ -45,7 +69,11 @@ public class RestaurantService {
         return this.repository.save(toUpdate);
     }
 
-    public void deleteRestaurant(final String id) {
+    public void deleteRestaurant(@NotNull final String id) {
+        if(StringUtils.isEmpty(id)){
+            throw new IllegalArgumentException("id cannot empty");
+        }
+
         this.repository.deleteById(UUID.fromString(id));
     }
 }
